requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.3            # via django
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
django-allauth==0.41.0
django-environ==0.4.5
django-extensions==2.2.6
django==3.0.2
idna==2.8                 # via requests
mysqlclient==1.4.6
oauthlib==3.1.0           # via requests-oauthlib
pillow==7.0.0
python3-openid==3.1.0     # via django-allauth
pytz==2019.3              # via django
requests-oauthlib==1.3.0  # via django-allauth
requests==2.22.0          # via django-allauth, requests-oauthlib
six==1.14.0               # via django-extensions
sqlparse==0.3.0           # via django
urllib3==1.25.8           # via requests
